% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_select_multiple_parent}
\alias{add_select_multiple_parent}
\title{Add select multiple parent column}
\usage{
add_select_multiple_parent(
  dt,
  parent_var,
  child_vars,
  choices,
  choices_sep = "/",
  sep = " ",
  na_value = NA
)
}
\arguments{
\item{dt}{(\code{data.table}) A data.table to add the parent column to.}

\item{parent_var}{(\code{character[1]}) Name of the new parent column to create.}

\item{child_vars}{(\verb{character[1+]}) Names of the child binary columns (0/1 values).}

\item{choices}{(\verb{character[1+]}) Labels/choices corresponding to each child column.}

\item{choices_sep}{(\code{character[1]}) Separator used to create well-named child variables.
child_vars will be renamed to paste0(parent_var, choices_sep, choices). (default: "/")}

\item{sep}{(\code{character[1]}) Separator to use when concatenating multiple selected options. (default: " ")}

\item{na_value}{(\code{logical[1] | character[1]}) Value to use when no options are selected. (default: NA)}
}
\value{
(\code{data.table}) The input data.table with the new parent column added.
}
\description{
Creates a parent column that concatenates the labels of selected options
from multiple binary columns (typically from select_multiple questions).
Can optionally create well-named child variables using parent_col and choices_sep.
}
\keyword{internal}
