% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_kobo_analysis.R
\name{auto_kobo_analysis}
\alias{auto_kobo_analysis}
\title{Auto survey from a Kobo tool analysis}
\usage{
auto_kobo_analysis(
  design,
  survey,
  choices,
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  choices_sep = "/",
  label_survey = TRUE,
  bind = TRUE
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{survey}{(\verb{df[1+, (name = character, type = character, ...)]}) A survey sheet.}

\item{choices}{(\verb{df[1+, (label = character, name = character, ...)]}) A choices sheet.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. Default to NULL for no group.}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{choices_sep}{(\code{character[1]}) Separator used in the Kobo extracted variable names for select_multiple questions. Default to "/".}

\item{label_survey}{(\code{logical[1]}) Whether to retrieve labels of vars from the survey sheet. Default to TRUE.}

\item{bind}{(\code{logical[1]}) Whether to output a data.table or a list. (default: TRUE for data.table)}
}
\value{
(``) A survey-summarized-median data frame
}
\description{
Auto survey from a Kobo tool analysis
}
\section{Automation}{


\link{auto_kobo_analysis} performs a very basic automated analysis using the provided Kobo tool.
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
