% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_svy_analysis.R
\name{auto_svy_analysis}
\alias{auto_svy_analysis}
\title{Auto survey analysis}
\usage{
auto_svy_analysis(
  design,
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  bind = TRUE,
  weight_var = NULL,
  stratum_var = NULL,
  cluster_var = NULL,
  id_var = NULL,
  fpc_var = NULL
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default: NULL for no group)}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}

\item{bind}{(``) Whether to output a data.table or a list. (default: TRUE for data.table)}

\item{weight_var}{(\verb{character[1+] | NULL}) Names of weight variables to exclude from analysis. (default: NULL)}

\item{stratum_var}{(\verb{character[1+] | NULL}) Names of stratum variables to exclude from analysis. (default: NULL)}

\item{cluster_var}{(\verb{character[1+] | NULL}) Names of cluster variables to exclude from analysis. (default: NULL)}

\item{id_var}{(\verb{character[1+] | NULL}) Names of ID variables to exclude from analysis. (default: NULL)}

\item{fpc_var}{(\verb{character[1+] | NULL}) Names of finite population correction variables to exclude from analysis. (default: NULL)}
}
\value{
(``) A survey-summarized-median data frame
}
\description{
Auto survey analysis
}
\section{Automation}{


\link{auto_svy_analysis} performs a very basic automated analysis. All numeric columns get means and medians. All character columns get a proportion.
}

\seealso{
Other survey analysis functions: 
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_proportion}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
