% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_missing_vars}
\alias{check_missing_vars}
\title{Check for missing variables in a data frame}
\usage{
check_missing_vars(df, vars, property = "vars", df_name = "df")
}
\arguments{
\item{df}{(\verb{data.frame[,1+]}) A data frame}

\item{vars}{(\verb{character[1+]}) Names of variables to check for in the data frame.}

\item{property}{(\code{character[1] | NULL}) Property being checked (default: 'vars').}

\item{df_name}{(\code{character[1]}) Name of the dataframe being checked, passed on to the abort message (default: 'df').}
}
\value{
(\code{TRUE | error}) TRUE if all variables are present, otherwise throws an error.
}
\description{
Check for missing variables in a data frame
}
\keyword{internal}
