% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{check_survey}
\alias{check_survey}
\title{Check if survey is minimally well-defined}
\usage{
check_survey(survey)
}
\arguments{
\item{survey}{(\verb{df[1+, (...)]}) A data frame from a survey sheet}
}
\description{
Check if survey is minimally well-defined
}
\keyword{internal}
