% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deselect}
\alias{deselect}
\title{Remove variable from a data frame}
\usage{
deselect(df, vars)
}
\arguments{
\item{df}{(\verb{df[, 1+]}) A data frame.}

\item{vars}{(\verb{character[1+]}) Variables to remove from \code{df}.}
}
\value{
(\code{dt}) A data.table with columns removed
}
\description{
Remove variable from a data frame
}
\keyword{internal}
