% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_diff}
\alias{df_diff}
\title{Remove variables from a data frame that exist in another dataframe}
\usage{
df_diff(df_a, df_b, vars_to_keep = NULL)
}
\arguments{
\item{df_a}{(\verb{df[, 1+]}) A data frame (or data.table) to remove variables from.}

\item{df_b}{(\verb{df[, 1+]}) A data frame (or data.table) that provides variable names to drop.}

\item{vars_to_keep}{(\verb{NULL | character[1+]}) Variables from \code{df_b} to keep in \code{df_a}. Supplying names here means "do \strong{not} drop these columns".}
}
\value{
(``) A data.table with the unwanted variables removed.
}
\description{
Remove variables from a data frame that exist in another dataframe
}
\keyword{internal}
