% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{get_survey_choices}
\alias{get_survey_choices}
\title{Get choices from survey name (and concatenate if you want!)}
\usage{
get_survey_choices(survey, choices, col, conc = TRUE, label = FALSE, sep = "_")
}
\arguments{
\item{survey}{(\verb{df[1+, (type = character, list_name = character, name = character, ...)]}) A survey sheet from Kobo (already split with columns list_name and name present).}

\item{choices}{(\verb{df[1+, (list_name = character, label = character, name = character, ...)]}) A choices sheet from Kobo (with column list_name, label and name).}

\item{col}{(\code{character[1]}) A column.}

\item{conc}{(\code{logical[1]}) Boolean. Should choices be concatenated to column name? Default to TRUE. Can only be used together with label = FALSE.}

\item{label}{(\code{logical[1]}) Whether labels be returned. It returns a data.frame with two columns: "name" and "label".}

\item{sep}{(\code{character[1]}) Separator for choices' concatenation.}
}
\value{
(\code{character | df}) A character vector of choices or pasted to \code{col} choices with a separator.
}
\description{
Get choices from survey name (and concatenate if you want!)
}
\keyword{internal}
