% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{get_survey_labels}
\alias{get_survey_labels}
\title{Get survey labels from survey names}
\usage{
get_survey_labels(survey, vars, output_df = TRUE)
}
\arguments{
\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column. See \code{split_survey()}.}

\item{vars}{(\verb{character[1+]}) Variable names.}

\item{output_df}{(\code{logical[1]}) Whether the output be a data frame (or a vector) Default to TRUE.}
}
\value{
(``) The labels as a vector.
}
\description{
Get survey labels from survey names
}
\keyword{internal}
