% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_analysis.R
\name{kobo_analysis}
\alias{kobo_analysis}
\title{Kobo survey analysis}
\usage{
kobo_analysis(
  design,
  analysis,
  vars,
  survey,
  choices = NULL,
  group = NULL,
  group_key_sep = " -/- ",
  auto_group_remove = TRUE,
  label_survey = TRUE,
  label_choices = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ratio_key_sep = " -/- ",
  interact_key_sep = " -/- ",
  choices_sep = "/"
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{analysis}{(\code{character[1]}) Analysis type. Either "mean", "median", "select_multiple", "select_one", "ratio", or "interact". See details for more, well, details.}

\item{vars}{(\verb{character[1+]}) Variables to calculate from.}

\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column. See \code{split_survey()}.}

\item{choices}{(\verb{NULL | df[1+, (label = character, name = character, ...)]}) The choices sheet from Kobo. If not NULL, the function tries to retrieve labels for choices.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default: NULL for no group)}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{auto_group_remove}{(\code{logical[1]}) If TRUE, the analysis won't be run for vars that are in group. Default to TRUE. (default: TRUE)}

\item{label_survey}{(\code{logical[1]}) Whether to retrieve labels of vars from the survey sheet. Default to TRUE.}

\item{label_choices}{(\code{logical[1]}) Whether to retrieve choices label from the choices sheet. Default to TRUE.}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{ratio_key_sep}{(\code{character[1]}) A separator for the ratio key. (default: " -/- ")}

\item{interact_key_sep}{(``) A character string to separate interactions columns in a fancy 'interact_key" column. (default: " -/- ")}

\item{choices_sep}{(\code{character[1]}) Separator used in the Kobo extracted variable names for select_multiple questions. Default to "/".}
}
\value{
(``) A survey-summarized-median data frame
}
\description{
Kobo survey analysis
}
\section{Specificity per type of analysis}{


The possible analysis type are: mean for \code{\link[=kobo_mean]{kobo_mean()}}, median for \code{\link[=kobo_median]{kobo_median()}}, select_one for \code{\link[=kobo_select_one]{kobo_select_one()}}, select_multiple for \code{\link[=kobo_select_multiple]{kobo_select_multiple()}}, or ratio for \code{\link[=kobo_ratio]{kobo_ratio()}}. It now also includes interactions between variables \code{\link[=kobo_interact]{kobo_interact()}} with interact#.
\itemize{
\item If \code{analysis} is "ratio": \code{vars} takes a named vector. Names will be passed to \code{nums} in \code{\link[=svy_ratio]{svy_ratio()}} and values to \code{denoms}.
\item If choices not NULL, the function tries to retrieve labels for select_one, while it is mandatory for select_multiple, and not used for mean, median, and ratio.
}
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
