% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_analysis_from_dap.R
\name{kobo_analysis_from_dap}
\alias{kobo_analysis_from_dap}
\alias{kobo_analysis_from_dap_group}
\title{Kobo survey analysis from a data analysis plan}
\usage{
kobo_analysis_from_dap(
  design,
  dap,
  survey,
  choices,
  group = NULL,
  level = 0.95,
  choices_sep = "/"
)

kobo_analysis_from_dap_group(
  design,
  dap,
  survey,
  choices,
  l_group,
  no_group = TRUE,
  level = 0.95,
  choices_sep = "/"
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{dap}{(\code{df}) A well formatted data analysis plan.}

\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column. See \code{split_survey()}.}

\item{choices}{(\verb{NULL | df[1+, (label = character, name = character, ...)]}) The choices sheet from Kobo. If not NULL, the function tries to retrieve labels for choices.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default: NULL for no group)}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{choices_sep}{(\code{character[1]}) Separator used in the Kobo extracted variable names for select_multiple questions. Default to "/".}

\item{l_group}{(\code{list[character]}) A list of vectors of variables to group by.}

\item{no_group}{(\code{logical[1]}) If TRUE, the analysis without grouping is run. (default: TRUE)}
}
\value{
(``) A survey-summarized-median data frame
}
\description{
\link{kobo_analysis_from_dap} performs a survey analysis from a data analysis plan, while \link{kobo_analysis_from_dap_group} is a wrapper around the former to run multiple grouped analyses at once.
}
\section{Specifics}{


A well formatted data analysis plan is a data frame with at least three columns: \code{analysis}, \code{vars}, \code{na_rm}, where,
\itemize{
\item \code{analysis} is one of the following types: "mean", "median", "select_multiple", "select_one", "ratio", "interact".
\item \code{vars} are the variables to analyze. For "mean", "median", "select_multiple", and "select_one", it is a single variable. For "ratio", it is two variables separated by a comma, where the first one is the numerator and the second one the denominator. For "interact", it is two or more variables separated by commas, which will be interacted.
\item \code{na_rm} is either "yes" or "no", indicating whether NAs should be removed for that analysis.
}
\itemize{
\item It takes the default for most variables, e.g. it tries to label, it calculates confidence intervals, and separators are the default.
\item Choices cannot be NULL.
\item The separator for choices can be changed depending on how your dataset's variable names are formatted.
\item Level can be changed.
}
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
