% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_interact.R
\name{kobo_interact}
\alias{kobo_interact}
\title{Get the proportion of the interaction of variables}
\usage{
kobo_interact(
  design,
  vars,
  interact_key_sep = " -/- ",
  survey,
  choices = NULL,
  group = NULL,
  group_key_sep = " -/- ",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{vars}{(\verb{character[1+]}) Variables to calculate interaction proportions from.}

\item{interact_key_sep}{(``) A character string to separate interactions columns in a fancy 'interact_key" column. (default: " -/- ")}

\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column. See \code{split_survey()}.}

\item{choices}{(\verb{NULL | df[1+, (label = character, name = character, ...)]}) The choices sheet from Kobo. If not NULL, the function tries to retrieve labels for choices. If not NULL, the function tries to retrieve labels.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default: NULL for no group)}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column.}

\item{label_survey}{(\code{logical[1]}) Whether to retrieve labels of vars from the survey sheet. (default: TRUE)}

\item{na_rm}{(\code{logical[1]}) Whether to remove NAs from \code{interact}. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}
}
\description{
A wrapper around \code{\link[=svy_interact]{svy_interact()}} to get the proportion of the interaction of variables, and using the Kobo tool to get labels.
}
