% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_median.R
\name{kobo_median}
\alias{kobo_median}
\title{Get the median for a numeric var}
\usage{
kobo_median(
  design,
  vars,
  survey,
  group = NULL,
  group_key_sep = " -/- ",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{vars}{(\verb{character[1+]}) Variables to calculate medians from.}

\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. Default to NULL for no group.}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{label_survey}{(\code{logical[1]}) Whether to retrieve labels of vars from the survey sheet. Default to TRUE.}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{NULL to report no variability. Otherwise one or more of: standard error ("se", the default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number indicating the confidence level (only one level allowed). Note that this may effect estimated standard errors (see \code{\link[survey]{svyquantile}} details on \code{alpha}, which equals \code{1-level}).}
}
\value{
(``) A character vector of select_one questions.
}
\description{
Get the median for a numeric var
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
