% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_ratio.R
\name{kobo_ratio}
\alias{kobo_ratio}
\title{Get the ratio for numeric variables.}
\usage{
kobo_ratio(
  design,
  nums,
  denoms,
  ratio_key_sep = " -/- ",
  survey,
  group = NULL,
  group_key_sep = " -/- ",
  label_survey = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{nums}{(\verb{character[1+]}) The numerator variables.}

\item{denoms}{(\verb{character[1+]}) The denominator variables.}

\item{ratio_key_sep}{(\code{character[1]}) A separator for the ratio key. (default: " -/- ")}

\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column. See \code{split_survey()}.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default: NULL for no group)}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column.}

\item{label_survey}{(\code{logical[1]}) Whether to retrieve labels of vars from the survey sheet. (default: TRUE)}

\item{na_rm}{(\code{logical[1]}) Whether to remove any line that as an NA in \code{num} or \code{denom} Default to TRUE.}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number or vector of numbers indicating the confidence level}
}
\value{
(``) A character vector of select_one questions.
}
\description{
Get the ratio for numeric variables.
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_select_multiple}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
