% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_select_multiple.R
\name{kobo_select_multiple}
\alias{kobo_select_multiple}
\title{Get the proportion for a select_multiple}
\usage{
kobo_select_multiple(
  design,
  vars,
  survey,
  choices = NULL,
  choices_sep = "/",
  group = NULL,
  group_key_sep = " -/- ",
  label_survey = TRUE,
  label_choices = TRUE,
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{vars}{(\verb{character[1+]}) Variables to calculate proportions from.}

\item{survey}{(\verb{df[1+, (type = character, name = character, ...)]}) The survey sheet from Kobo with column "type" split already, see \code{?split_survey()}. If \code{label_survey} is TRUE, survey must also contain a "label" column. See \code{split_survey()}.}

\item{choices}{(\verb{NULL | df[1+, (label = character, name = character, ...)]}) The choices sheet from Kobo. If not NULL, the function tries to retrieve labels for choices.}

\item{choices_sep}{(\code{character[1]}) Separator used in the Kobo extracted variable names for select_multiple questions. Default to "/".}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. Default to NULL for no group.}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{label_survey}{(\code{logical[1]}) Whether to retrieve labels of vars from the survey sheet. Default to TRUE.}

\item{label_choices}{(\code{logical[1]}) Whether to retrieve choices label from the choices sheet. Default to TRUE.}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}
}
\value{
(``) A character vector of select_one questions.
}
\description{
Get the proportion for a select_multiple
}
\details{
\code{survey} should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
\section{Not removing missing values}{


The rationale when not removing missing values is the following:
\itemize{
\item Missing values for each dummy 1/0 column corresponding to response option are recoded to 0
\item Then, the mean of 0s and 1s is computed, thus the \% for all response options
}

This allows to calculate the \% for each choice over the whole dataset.
}

\seealso{
Other functions for analyzing from Kobo tool: 
\code{\link{auto_kobo_analysis}()},
\code{\link{kobo_analysis}()},
\code{\link{kobo_analysis_from_dap}()},
\code{\link{kobo_mean}()},
\code{\link{kobo_mean_all}()},
\code{\link{kobo_median}()},
\code{\link{kobo_median_all}()},
\code{\link{kobo_ratio}()},
\code{\link{kobo_select_multiple_all}()},
\code{\link{kobo_select_one}()},
\code{\link{kobo_select_one_all}()}
}
\concept{functions for analyzing from Kobo tool}
