% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{named_group_split}
\alias{named_group_split}
\title{Named group split}
\usage{
named_group_split(df, group, warn_if_set_dt = FALSE)
}
\arguments{
\item{df}{(\verb{df[, 1+]}) A data frame.}

\item{group}{(\code{character[1]}) Variable name to group by.}

\item{warn_if_set_dt}{(\code{logical[1]}) Whether to warn if df is coerced to data.table. (default: FALSE)}
}
\value{
(``) A split and named list of data.frames/data.tables.
}
\description{
Named group split
}
\keyword{internal}
