% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_order_before}
\alias{set_order_before}
\title{Create column order to insert new column before target column}
\usage{
set_order_before(df, vars, target_var, warn_if_unchanged = FALSE)
}
\arguments{
\item{df}{(\code{data.frame}) A data frame or data.table.}

\item{vars}{(\code{character[1]}) Name of the variables to move.}

\item{target_var}{(\code{character[1]}) Name of the target column that the variables should be moved before.}

\item{warn_if_unchanged}{(\code{logical[1]}) Whether to warn if the columns are already in the right place. (default: FALSE)}
}
\value{
(\code{character}) A character vector with the new column order.
}
\description{
Create column order to insert new column before target column
}
\keyword{internal}
