% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_utils.R
\name{split_survey}
\alias{split_survey}
\title{Split survey type and list name}
\usage{
split_survey(
  survey,
  col_to_split,
  into = c("type", "list_name"),
  sep = " ",
  fill = "right",
  ...
)
}
\arguments{
\item{survey}{(\verb{df[1+, (...)]}) A survey sheet from Kobo}

\item{col_to_split}{(\code{character[1]}) Usually \code{type}}

\item{into}{(\verb{character[1+]}) Vector of columns names to split to. Default to c("type", "list_name" )}

\item{sep}{(\code{character[1]}) The separator. Default to " ".}

\item{fill}{(\code{character[1]}) How to fill. Default to NA on the right.}

\item{...}{Params to pass to \code{tidyr::separate()}.}
}
\value{
(``) A survey data.table, split
}
\description{
Split survey type and list name
}
\keyword{internal}
