% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveyLiteR-package.R
\docType{package}
\name{SurveyLiteR-package}
\alias{SurveyLiteR}
\alias{SurveyLiteR-package}
\title{SurveyLiteR: Analyzing survey data (collected with Kobo)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Providing helpers to support a (quasi-)automated survey data analysis workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gnoblet/SurveyLiteR}
}

}
\author{
\strong{Maintainer}: Noblet Guillaume \email{gnoblet@zaclys.net}

}
\keyword{internal}
