% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_interact.R
\name{svy_interact}
\alias{svy_interact}
\title{Survey interactions}
\usage{
svy_interact(
  design,
  interact,
  interact_key_sep = " -/- ",
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ak = TRUE,
  ak_overall_sep = " @/@ ",
  ak_main_sep = " -/- ",
  ak_var_to_value_sep = " \%/\% ",
  ...
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{interact}{(\verb{character[2+]}) Variables to calculate interactions from.}

\item{interact_key_sep}{(``) A character string to separate interactions columns in a fancy 'interact_key" column. (default: " -/- ")}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. Default to NULL for no group.}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column.}

\item{na_rm}{(\code{logical[1]}) Whether to remove NAs from \code{interact}. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{ak}{(\code{logical[1]}) Whether to add an analysis key column. (default: TRUE)}

\item{ak_overall_sep}{(\code{character[1]}) The overall separator between items, e.g. between the type of analysis and the variables information.}

\item{ak_main_sep}{(\code{character[1]}) The main separator between variables, e.g. between the two grouping columns.}

\item{ak_var_to_value_sep}{(\code{character[1]}) The separator between the variable and its value.}

\item{...}{Other parameters to pass to \code{srvyr::survey_mean()}.}
}
\value{
(``) A survey-summarized-mean data frame
}
\description{
Survey interactions
}
\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_analysis}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_proportion}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
