% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_make.R
\name{svy_make_ratio}
\alias{svy_make_ratio}
\title{Helper function for ratio calculations}
\usage{
svy_make_ratio(
  design,
  nums,
  denoms,
  ratio_key_sep = " -/- ",
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{A srvyr::design object}

\item{nums}{Character vector of numerator variables}

\item{denoms}{Character vector of denominator variables}

\item{ratio_key_sep}{Separator for ratio keys}

\item{group}{Character vector of grouping variables}

\item{group_key_sep}{Separator for group key values}

\item{na_rm}{Whether to remove NAs}

\item{vartype}{Type of variance estimate}

\item{level}{Confidence level}

\item{...}{Additional arguments passed to srvyr::survey_ratio}
}
\value{
A data.table with survey ratio results
}
\description{
Special helper for svy_ratio since it works with pairs of variables
}
\keyword{internal}
