% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_proportion.R
\name{svy_proportion}
\alias{svy_proportion}
\title{Survey proportion}
\usage{
svy_proportion(
  design,
  vars,
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ak = TRUE,
  ak_overall_sep = " @/@ ",
  ak_main_sep = " -/- ",
  ak_var_to_value_sep = " \%/\% ",
  ...
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{vars}{(\verb{character[1+]}) Variables to calculate the proportoins from.}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default to NULL for no group)}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{Report variability as one or more of: standard error ("se", default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{(For vartype = "ci" only) A single number or vector of numbers indicating
the confidence level}

\item{ak}{(\code{logical[1]}) Whether to add an analysis key column. (default: TRUE)}

\item{ak_overall_sep}{(\code{character[1]}) The overall separator between items, e.g. between the type of analysis and the variables information. (default: " @/@ ")}

\item{ak_main_sep}{(\code{character[1]}) The main separator between variables, e.g. between the two grouping columns. (default: " -/- ")}

\item{ak_var_to_value_sep}{(\code{character[1]}) The separator between the variable and its value. (default: " \%/\% ")}

\item{...}{Other parameters to pass to \code{srvyr::survey_prop()}.}
}
\value{
(``) A survey-summarized-proportion data frame
}
\description{
Survey proportion
}
\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_quantile}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
