% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_quantile.R
\name{svy_quantile}
\alias{svy_quantile}
\title{Survey quantile}
\usage{
svy_quantile(
  design,
  vars,
  quantiles = c(0.25, 0.5, 0.75),
  group = NULL,
  group_key_sep = " -/- ",
  na_rm = TRUE,
  vartype = "ci",
  level = 0.95,
  ...
)
}
\arguments{
\item{design}{(\code{srvyr::design}) A srvyr::design object.}

\item{vars}{(\verb{character[1+]}) Variables to calculate quantiles from.}

\item{quantiles}{A vector of quantiles to calculate}

\item{group}{(\verb{character[1+] | NULL}) Variables to group by. (default: NULL for no group)}

\item{group_key_sep}{(\code{character[1]}) The string to separate grouping variables in a fancy 'group_key' column. (default: " -/- ")}

\item{na_rm}{(\code{logical[1]}) Whether NAs in 'var' should be removed. (default: TRUE)}

\item{vartype}{NULL to report no variability. Otherwise one or more of: standard error ("se", the default),
confidence interval ("ci"), variance ("var") or coefficient of variation
("cv").}

\item{level}{A single number indicating the confidence level (only one level allowed). Note that this may effect estimated standard errors (see \code{\link[survey]{svyquantile}} details on \code{alpha}, which equals \code{1-level}).}

\item{...}{Other parameters to pass to \code{srvyr::survey_quantile()}.}
}
\value{
(\code{data.frame}) A survey-summarized-mean data frame
}
\description{
Survey quantile
}
\seealso{
Other survey analysis functions: 
\code{\link{auto_svy_analysis}()},
\code{\link{svy_analysis}()},
\code{\link{svy_interact}()},
\code{\link{svy_mean}()},
\code{\link{svy_median}()},
\code{\link{svy_proportion}()},
\code{\link{svy_ratio}()}
}
\concept{survey analysis functions}
